Reader.ts -> Argument.ts & translator stages renamed
