+++ src/command/argument/index.ts
@@ -1,4 +1,3 @@
 export * from "./Argument";
-export * from "./Reader";
 export * from "./Sequence";
 export * from "./readers";
