+++ src/command/argument/readers/Mention.ts
@@ -1,5 +1,5 @@
 import { CommandContext } from "../../Context";
-import { ArgumentReader } from "../Reader";
+import { ArgumentReader } from "../Argument";
 import { parsedRegexReader } from "./Regex";
 
 /**
